+++ pwa/src/pages/translations/[id].tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 import Layout from "../../components/common/layout";
 import TranslationForm from "../../components/translations/translationForm";
 import { Tabs } from "@conductionnl/nl-design-system/lib/Tabs/src/tabs";
+// import TranslationTable from "../../components/translations/translationTable";
 
 const IndexPage = (props) => {
   const [context, setContext] = React.useState(null);
@@ -20,17 +21,12 @@ const IndexPage = (props) => {
         <div className="row">
           <div className="col-12">
             <div className="page-top-item">
-              {props.params.id !== "new" ? (
                 <Tabs
                   items={[
-                    { name: "Overview", id: "overview", active: true }
+                    { name: "Add", id: "overview", active: true },
+                    { name: "Translations", id: "translations"}
                   ]}
                 />
-              ) : (
-                <Tabs
-                  items={[{ name: "Overview", id: "overview", active: true }]}
-                />
-              )}
             </div>
             <div className="tab-content">
               <div
@@ -41,6 +37,16 @@ const IndexPage = (props) => {
               >
                 <br />
                 <TranslationForm id={props.params.id} />
+              </div>
+              <div
+                className="tab-pane"
+                id="translations"
+                role="tabpanel"
+                aria-labelledby="translations-tab"
+              >
+                <br />
+
+               {/*<TranslationTable id={props.params.id}/>*/}
               </div>
             </div>
           </div>
